task_handler/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
Click==7.0
Flask==1.1.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.5
MarkupSafe==1.1.1
more-itertools==7.2.0
pytz==2019.3
redis==3.3.8
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
openslide-python==1.1.1
opencv-python==3.4.4.19
Pillow==6.2.0
scipy==1.1.0 
scikit-image==0.14.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
requests==2.22.0